\name{checkShow}
\alias{checkShow}
\title{
Show the current checks executed in the TRES DB
}
\description{
Reads the \code{\link{checkBD}} function and extracts the different checks that will be returned as a data.frame or an html table.
}
\usage{
checkShow(html = FALSE)
}
\arguments{
  \item{html}{Logical}
}
\details{}
\value{
A data.frame or an html table
}
\author{
Francois Rousseu, Nghia Tran and Audrey Bourret
}
\seealso{
\code{\link{checkBD}}, \code{\link{htmlTable}}
}
\examples{
}

