\name{print.TREScheck}
\alias{print.TREScheck}
\title{
Print method for a \code{TREScheck} list of checks
}
\description{
Prints a list with the check results
}
\usage{
## S3 method for class 'TREScheck'
print(x)
x
}
\arguments{
  \item{x}{The list returned by \code{\link{checkBD}}.}
}
\details{
 The list can also be printed just by typing the name of the object in the console.
}
\value{
A list printed in the console.
}
\author{
Francois Rousseu, Nghia Tran and Audrey Bourret
}
\seealso{
\code{\link{checkBD}}
}
\examples{}

